refactor(registers): atualiza importacoes de estilos para novos cards de registro
