+++ src/views/pages/dashboard/registers/edit/components/PatchableRegisterCard/index.tsx
@@ -6,11 +6,11 @@ import {
 	ReceiptIcon,
 	ReceiptXIcon,
 } from '@phosphor-icons/react';
-import { RegisterItem } from '../RegisterItem';
-import { Container, HeadText, Total, Values } from '../RegisterCard/styles';
 import { Controller, useFormContext } from 'react-hook-form';
 import { calculateRegisterTotal } from '@/utils/calculate-register-total';
 import { EditRegisterSchema } from '@/schemas/edit-register-schema';
+import { Container, HeadText, Total, Values } from '../../../components/Cards/RegisterCard/styles';
+import { RegisterItem } from '../../../components/Cards/RegisterItem';
 
 interface PatchableRegisterCardProps {
 	id: number;
